pyproject.toml
[project]
name = "rlm-implementation"
version = "0.1.0"
description = "A simplified Recursive Language Model (RLM) demo with true context isolation."
requires-python = ">=3.10"
dependencies = [
    "litellm>=1.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rlm"]
